+++ b2sdk/_v2/__init__.py
@@ -198,6 +198,8 @@ from b2sdk.sync.encryption_provider import SERVER_DEFAULT_SYNC_ENCRYPTION_SETTIN
 # other
 
 from b2sdk.b2http import B2Http
+from b2sdk.api_config import B2HttpApiConfig
+from b2sdk.api_config import DEFAULT_HTTP_API_CONFIG
 from b2sdk.b2http import ClockSkewHook
 from b2sdk.b2http import HttpCallback
 from b2sdk.b2http import ResponseContextManager