+++ b2sdk/v1/__init__.py
@@ -13,6 +13,7 @@ from b2sdk.v1.account_info import (
     AbstractAccountInfo, InMemoryAccountInfo, UrlPoolAccountInfo, SqliteAccountInfo, StubAccountInfo
 )
 from b2sdk.v1.api import B2Api
+from b2sdk.v1.b2http import B2Http
 from b2sdk.v1.bucket import Bucket, BucketFactory
 from b2sdk.v1.cache import AbstractCache
 from b2sdk.v1.download_dest import (
@@ -28,3 +29,5 @@ from b2sdk.v1.sync import (
     LocalFolder, B2Folder, parse_sync_folder, File, B2File, FileVersion,
     AbstractSyncEncryptionSettingsProvider
 )
+
+B2RawApi = B2RawHTTPApi