+++ test/unit/b2http/test_b2http.py
@@ -15,9 +15,11 @@ import sys
 
 from ..test_base import TestBase
 
+import apiver_deps
 from apiver_deps_exception import BadDateFormat, BadJson, BrokenPipe, B2ConnectionError, ClockSkew, ConnectionReset, ServiceError, UnknownError, UnknownHost, TooManyRequests
 from apiver_deps import USER_AGENT
 from apiver_deps import B2Http
+from apiver_deps import B2HttpApiConfig
 from apiver_deps import ClockSkewHook
 from apiver_deps import translate_errors as _translate_errors
 from apiver_deps import translate_and_retry as _translate_and_retry
@@ -210,9 +212,17 @@ class TestB2Http(TestBase):
 
         requests = MagicMock()
         requests.Session.return_value = self.session
-        self.b2_http = B2Http(
-            requests, install_clock_skew_hook=False, user_agent_append=self.UA_APPEND
-        )
+
+        if apiver_deps.V <= 1:
+            self.b2_http = B2Http(
+                requests, install_clock_skew_hook=False, user_agent_append=self.UA_APPEND
+            )
+        else:
+            self.b2_http = B2Http(
+                B2HttpApiConfig(
+                    requests, install_clock_skew_hook=False, user_agent_append=self.UA_APPEND
+                )
+            )
 
     def test_post_json_return_json(self):
         self.session.post.return_value = self.response