+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `B2Api.get_file_info` returns a `FileVersion` object in v2
 * `B2RawApi` renamed to `B2RawHTTPApi`
 * `B2HTTP` tests are now common
+* `B2HttpApiConfig` class introduced to provide parameters like `user_agent_append` to `B2Api` without using internal classes in v2
 
 ## [1.9.0] - 2021-06-07